backend/requirements.txt
# Core Django
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# PostGIS & Spatial
django-crispy-forms==2.1

# Redis & Celery
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.6.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-auditlog==3.0.0
django-log-request-id==2.1.0

# WebSockets
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Logging & Monitoring
structlog==24.1.0
django-prometheus==2.3.1
python-json-logger==2.0.7

# AI & ML (lightweight)
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0

# Geospatial
geopy==2.4.1
shapely==2.0.2

# API Utilities
django-filter==23.5
drf-yasg==1.21.7

# Development
python-dotenv==1.0.0
setuptools>=65.0.0

# TimescaleDB support
django-timescaledb==0.2.3

